Use assertions to check workflow
Use Java `assert` to check whether componenets with dependencies are initialized in a proper order.
