+++ build.gradle
@@ -18,6 +18,7 @@ dependencies {
     testRuntimeOnly group: 'org.junit.jupiter', name: 'junit-jupiter-engine', version: '5.5.0'
 }
 dependencies {
+    implementation 'org.jetbrains:annotations:20.1.0'
     implementation 'org.testng:testng:7.1.0'
     implementation 'junit:junit:4.13.1'
     testImplementation 'junit:junit:4.12'
@@ -53,4 +54,8 @@ tasks.register('copyTaskWithPatterns', Copy) {
     from '$rootDir'
     into layout.buildDirectory.dir('')
     include '**/*.jpg'
+}
+
+build {
+    shadowJar
 }
