+++ src/main/java/org/Olivia/IO/DialogBox.java
@@ -1,4 +1,4 @@
-package org.Olivia;
+package org.Olivia.IO;
 
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
