+++ src/main/java/org/Olivia/Olivia.java
@@ -1,174 +1,45 @@
 package org.Olivia;
 
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
-import javafx.geometry.Pos;
-import javafx.scene.Node;
-import javafx.scene.control.Label;
-import javafx.scene.layout.Region;
-import org.Olivia.IO.FileHandler;
-import org.Olivia.IO.UiHandler;
+import org.Olivia.Dispatchers.CliEventDispatcher;
+import org.Olivia.IO.*;
 import org.Olivia.calendar.Calendar;
 import javafx.application.Application;
-import javafx.scene.Scene;
-import javafx.scene.control.Button;
-import javafx.scene.control.ScrollPane;
-import javafx.scene.control.TextField;
-import javafx.scene.layout.AnchorPane;
-import javafx.scene.layout.VBox;
 import javafx.stage.Stage;
-import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
 
-import java.io.File;
-import java.io.InputStream;
-import java.io.FileInputStream;
-
+/**
+ * This is the application entry point
+ * In the GUI version, the JavaFX launcher will call the start function in this class to setup the app
+ * In the CLI version (obsolete), call the main function in this class to start the application
+ * @author ZHANG TONGJUN (albertZhangTJ)
+ */
 
 public class Olivia extends Application {
 
-    private ScrollPane scrollPane;
-    private VBox dialogContainer;
-    private TextField userInput;
-    private Button sendButton;
-    private Scene scene;
-    private Image user;// = new Image(Objects.requireNonNull(this.getClass().getResourceAsStream("/images/User.jpg")));
-    private Image olivia;// = new Image(this.getClass().getResourceAsStream("/images/Olivia.jpg"));
-    private GuiEventDispatcher dispatcher;
+    private Gui UI;
 
 
+    @Deprecated
+    /**
+     * This main method is the entry point for the CLI version of the software
+     * As the GUI is already implemented, this is no longer needed
+     * However, it is kept here in case someone want to try with the CLI version
+     */
     public static void main(String[] args) {
         //each worker is attached to a calendar table to work on and an UiHandler to handle IO
         //this design would allow the program to handle multi-user with multi-input-source case (with multi-thread)
         //the UiHandler can be binded to a specific InputStream and a specific PrintStream (default to be stdin & stdout)
-        EventDispatcher worker_1 = new EventDispatcher(new Calendar(), new UiHandler(), new FileHandler());
+        CliEventDispatcher worker_1 = new CliEventDispatcher(new Calendar(), new UiHandler(), new FileHandler());
         worker_1.startWorking();
     }
 
-    //@Override
+    @Override
     public void start(Stage stage) {
-        // somehow the getResourceAsStream function does not work on my system
-        // (it always returns null)
-        try {
-//            File userIconFile = new File("images/User.jpg");
-//            assert userIconFile!=null;
-//            InputStream userIconStream = new FileInputStream(userIconFile);
-//            File oliviaIconFile = new File("images/Olivia.jpg");
-//            assert oliviaIconFile!=null;
-//            InputStream oliviaIconStream = new FileInputStream(oliviaIconFile);
-            InputStream userIco=getClass().getResourceAsStream("/images/User.jpg");
-            InputStream oliviaIco=getClass().getResourceAsStream("/images/Olivia.jpg");
-            user = new Image(userIco);
-            olivia = new Image(oliviaIco);
-        }
-        catch (Exception e) {
-            e.printStackTrace();
-        }
-        this.dispatcher=new GuiEventDispatcher(new Calendar(), new FileHandler());
-        try {
-            this.dispatcher.initialize();
-        }
-        catch (Throwable e) {
-            System.out.println(e.getMessage());
-        }
-        //The container for the content of the chat to scroll.
-        scrollPane = new ScrollPane();
-        dialogContainer = new VBox();
-        scrollPane.setContent(dialogContainer);
 
-        userInput = new TextField();
-        sendButton = new Button("Send");
-
-        AnchorPane mainLayout = new AnchorPane();
-        mainLayout.getChildren().addAll(scrollPane, userInput, sendButton);
-
-        scene = new Scene(mainLayout);
-
-        stage.setScene(scene);
-        stage.show();
-
-        stage.setTitle("Duke");
-        stage.setResizable(false);
-        stage.setMinHeight(600.0);
-        stage.setMinWidth(400.0);
-
-        mainLayout.setPrefSize(400.0, 600.0);
-
-        scrollPane.setPrefSize(385, 535);
-        scrollPane.setHbarPolicy(ScrollPane.ScrollBarPolicy.NEVER);
-        scrollPane.setVbarPolicy(ScrollPane.ScrollBarPolicy.ALWAYS);
-
-        scrollPane.setVvalue(1.0);
-        scrollPane.setFitToWidth(true);
-
-        // You will need to import `javafx.scene.layout.Region` for this.
-        dialogContainer.setPrefHeight(Region.USE_COMPUTED_SIZE);
-
-        userInput.setPrefWidth(325.0);
-
-        sendButton.setPrefWidth(55.0);
-
-        AnchorPane.setTopAnchor(scrollPane, 1.0);
-
-        AnchorPane.setBottomAnchor(sendButton, 1.0);
-        AnchorPane.setRightAnchor(sendButton, 1.0);
-
-        AnchorPane.setLeftAnchor(userInput, 1.0);
-        AnchorPane.setBottomAnchor(userInput, 1.0);
-        dialogContainer.heightProperty().addListener((observable) -> scrollPane.setVvalue(1.0));
-
-
-        sendButton.setOnMouseClicked((event) -> {
-            handleUserInput();
-        });
-
-        userInput.setOnAction((event) -> {
-            handleUserInput();
-        });
+        this.UI = new Gui(stage, GuiApplicationInitializer.setUpGuiEventDispatcher());
+        this.UI.initialize();
     }
 
-    /**
-     * Iteration 1:
-     * Creates a label with the specified text and adds it to the dialog container.
-     *
-     * @param text String containing text to add
-     * @return a label with the specified text that has word wrap enabled.
-     */
-    private Label getDialogLabel(String text) {
-        // You will need to import `javafx.scene.control.Label`.
-        Label textToAdd = new Label(text);
-        textToAdd.setWrapText(true);
 
-        return textToAdd;
-    }
-
-    /**
-     * Iteration 2:
-     * Creates two dialog boxes, one echoing user input and the other containing Duke's reply and then appends them to
-     * the dialog container. Clears the user input after processing.
-     */
-    private void handleUserInput() {
-        Label userText = new Label(userInput.getText());
-        Label dukeText = new Label(getResponse(userInput.getText()));
-        dialogContainer.getChildren().addAll(
-                DialogBox.getUserDialog(userText, new ImageView(user)),
-                DialogBox.getOliviaDialog(dukeText, new ImageView(olivia))
-        );
-        userInput.clear();
-    }
-
-    /**
-     * You should have your own function to generate a response to user input.
-     * Replace this stub with your completed method.
-     */
-    private String getResponse(String input) {
-        try {
-            return dispatcher.dispatchCommand(input);
-        }
-        catch (Exception e){
-            return e.getMessage();
-        }
-    }
 
 
 }
